bigger and badder fix

+++ scripts/RunScript.hx
@@ -138,7 +138,11 @@ class RunScript {
         var cacheDirectory:Array<String> = FileSystem.readDirectory("shaders/VKGL");
 
         for(i in 0...cacheDirectory.length) {
-            FileSystem.deleteFile("shaders/VKGL/" + cacheDirectory[i]);
+            var split:Array<String> = cacheDirectory[i].split(".");
+
+            if(split[split.length - 1] == "spv") {
+                FileSystem.deleteFile("shaders/VKGL/" + cacheDirectory[i]);
+            }
         }
 
         for(i in 0...cacheDirectory.length) {
